refactor: apply clean architecture
resolves: #20
