+++ src/main/kotlin/com/kroffle/knitting/data/entity/design/DesignType.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.domain
+package com.kroffle.knitting.data.entity.design
 
 enum class DesignType(val code: Int) {
     Sweater(1),
