+++ src/main/kotlin/com/kroffle/knitting/data/entity/design/PatternType.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.domain
+package com.kroffle.knitting.data.entity.design
 
 enum class PatternType(val code: Int) {
     Text(1),
