+++ src/main/kotlin/com/kroffle/knitting/data/repository/design/DesignRepository.kt
@@ -1,0 +1,6 @@
+package com.kroffle.knitting.data.repository.design
+
+import com.kroffle.knitting.data.entity.design.DesignEntity
+import org.springframework.data.repository.reactive.ReactiveCrudRepository
+
+interface DesignRepository : ReactiveCrudRepository<DesignEntity, Long>
