--- src/main/kotlin/com/kroffle/knitting/domain/DesignRepository.kt
@@ -1,5 +1,0 @@
-package com.kroffle.knitting.domain
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository
-
-interface DesignRepository : ReactiveCrudRepository<Design, Long>
